버그 수정 -유저 모델이 Save될때마다 비밀번호가 hash되는 오류 수정 -Video edit/delete 접근권한 부여

+++ src/controllers/videoController.js
@@ -18,18 +18,27 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id: userId } } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." })
+    }
+    if(String(video.owner) !== String(userId)){
+        return res.status(403).redirect("/");
     }
     return res.render("videos/edit", { pageTitle: `Edit ${video.title}`, video });
 }
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id: userId } } = req.session;
     const { title, description, hashtags } = req.body;
-    if (!(await Video.exists({ _id: id }))) {
+    const video = await Video.findById(id);
+    if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." })
+    }
+    if(String(video.owner) !== String(userId)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -41,9 +50,19 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {
+      user: { _id },
+    } = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+      return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
-    return res.redirect(`/`);
-}
+    return res.redirect("/");
+  };
 
 export const getUpload = (req, res) => {
     return res.render("videos/upload", { pageTitle: "Upload Video" });
